refactor: split AccessibilityPage and PrivacyPolicyPage into sub-components
Per reviewer feedback (@maxgfr), split large single-file pages into
folder-based modules with logically grouped sub-components:

AccessibilityPage/ (was 206 lines)
  - index.tsx: page orchestrator
  - AccessibilityCompliance.tsx: status, test results, non-accessible content
  - AccessibilityEstablishment.tsx: dates, technologies, testing environment
  - AccessibilityContact.tsx: feedback and recourse information

PrivacyPolicyPage/ (was 333 lines)
  - index.tsx: page orchestrator
  - PrivacyOverview.tsx: responsible entity, data processed, purpose, legal basis
  - PrivacyRightsAndData.tsx: retention, rights, access, subprocessors
  - PrivacyCookies.tsx: cookie policy and CNIL references
