+++ packages/app/src/modules/legal/PrivacyPolicyPage/PrivacyOverview.tsx
@@ -1,0 +1,51 @@
+/** Responsible entity, data processed, purpose, and legal basis. */
+export function PrivacyOverview() {
+	return (
+		<>
+			<h2 className="fr-h4">Qui est responsable d'Index Egapro ?</h2>
+			<p>
+				La plateforme Index Egapro relève de la Direction générale du travail,
+				représentée par Monsieur Pierre RAMAIN, directeur général du travail.
+				Index Egapro met en place deux index conçus pour lutter contre
+				l'inégalité entre les femmes et les hommes.
+			</p>
+
+			<h2 className="fr-h4">
+				Quelles sont les données à caractère personnel que nous traitons ?
+			</h2>
+			<p>Index Egapro traite les données à caractère personnel suivantes :</p>
+			<ul>
+				<li>
+					Nom, prénom, adresse e-mail et numéro de téléphone professionnel des
+					utilisateurs.
+				</li>
+			</ul>
+
+			<h2 className="fr-h4">
+				Pourquoi traitons-nous des données à caractère personnel ?
+			</h2>
+			<p>Index Egapro traite des données à caractère personnel pour :</p>
+			<ul>
+				<li>
+					Calculer les écarts de rémunération entre les femmes et les hommes au
+					sein des entreprises et réaliser les déclarations qui en sont issues ;
+				</li>
+				<li>
+					Calculer les écarts de représentation équilibrée entre les femmes et
+					les hommes dans les services dirigeants.
+				</li>
+			</ul>
+
+			<h2 className="fr-h4">
+				Qu'est-ce qui nous autorise à traiter des données à caractère personnel
+				?
+			</h2>
+			<p>
+				Index Egapro traite des données à caractère personnel en se basant sur
+				l'exécution d'une mission d'intérêt public ou relevant de l'exercice de
+				l'autorité publique dont est investi le responsable de traitement
+				conformément à l'article 6-1 e) du RGPD.
+			</p>
+		</>
+	);
+}
